requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
openai==2.7.2
pdf2image==1.17.0
pillow==11.3.0
python-dotenv==1.2.1
reportlab==4.4.4
Werkzeug==3.1.3
gunicorn==23.0.0
